+++ release-notes.md
@@ -1,3 +1,11 @@
+v 0.13.4
+========
+
+- added `Modalities.get_all_ids()`
+- added `Modalities.get_id_from_aet()`
+- added `Study.patient_orthanc_id`
+- added `Resources.exists()`
+
 v 0.13.3
 ========